+++ CHANGELOG.md
@@ -10,6 +10,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Changed
 
+## [0.0.5] - 2026-08-12
+
+### Added
+
+- REST API endpoints to fetch agent conversations and message history.
+- Automatic creation and assignment of conversations to agents in `MessagingService`.
+- `agent_id` field to `conversations` database schema. (flask db upgrade)
+
 ## [0.0.4] - 2026-08-11 
 
 ### Added
@@ -51,7 +59,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - error handling and debug utils
 - llm service layer vor gemini
 
-[Unreleased]: https://github.com/negsi/trinity-flask/compare/v0.0.4...develop
+[Unreleased]: https://github.com/negsi/trinity-flask/compare/v0.0.5...develop
+[0.0.5]: https://github.com/negsi/trinity-flask/releases/tag/v0.0.5
 [0.0.4]: https://github.com/negsi/trinity-flask/releases/tag/v0.0.4
 [0.0.3]: https://github.com/negsi/trinity-flask/releases/tag/v0.0.3
 [0.0.2]: https://github.com/negsi/trinity-flask/releases/tag/v0.0.2